+++ tests/mod.rs
@@ -6,6 +6,7 @@
 use std::env;
 
 mod detect_test;
+mod django_test;
 mod package_manager_test;
 mod pip_test;
 mod python_version_test;